refactor into separate files for api and net
